+++ noodles-bcf/src/io/indexed_reader.rs
@@ -4,7 +4,7 @@ mod builder;
 
 pub use self::builder::Builder;
 
-use std::io::{self, Read, Seek};
+use std::io::{self, Read};
 
 use noodles_bgzf as bgzf;
 use noodles_core::Region;
@@ -98,9 +98,9 @@ where
     }
 }
 
-impl<R> IndexedReader<bgzf::Reader<R>>
+impl<R> IndexedReader<R>
 where
-    R: Read + Seek,
+    R: bgzf::io::BufRead + bgzf::io::Seek,
 {
     /// Returns an iterator over records that intersects the given region.
     pub fn query<'r, 'h>(