+++ noodles-bcf/src/io/reader/query.rs
@@ -1,4 +1,4 @@
-use std::io::{self, Read, Seek};
+use std::io;
 
 use noodles_bgzf as bgzf;
 use noodles_core::region::Interval;
@@ -11,10 +11,7 @@ use crate::Record;
 /// An iterator over records of a BCF reader that intersects a given region.
 ///
 /// This is created by calling [`super::Reader::query`].
-pub struct Query<'r, 'h, R>
-where
-    R: Read + Seek,
-{
+pub struct Query<'r, 'h, R> {
     reader: csi::io::Query<'r, R>,
     header: &'h vcf::Header,
     chromosome_id: usize,
@@ -24,10 +21,10 @@ where
 
 impl<'r, 'h, R> Query<'r, 'h, R>
 where
-    R: Read + Seek,
+    R: bgzf::io::BufRead + bgzf::io::Seek,
 {
     pub(super) fn new(
-        reader: &'r mut bgzf::Reader<R>,
+        reader: &'r mut R,
         header: &'h vcf::Header,
         chunks: Vec<Chunk>,
         chromosome_id: usize,
@@ -52,7 +49,7 @@ where
 
 impl<'r, 'h, R> Iterator for Query<'r, 'h, R>
 where
-    R: Read + Seek,
+    R: bgzf::io::BufRead + bgzf::io::Seek,
 {
     type Item = io::Result<Record>;