+++ noodles-sam/src/io/reader.rs
@@ -8,7 +8,7 @@ pub(crate) mod record_buf;
 mod record_bufs;
 
 use std::{
-    io::{self, BufRead, Read, Seek},
+    io::{self, BufRead},
     iter,
 };
 
@@ -276,15 +276,18 @@ where
     }
 }
 
-impl<R> Reader<bgzf::Reader<R>>
+impl<R> Reader<R>
 where
-    R: Read + Seek,
+    R: bgzf::io::BufRead + bgzf::io::Seek,
 {
     // Seeks to the first record by setting the cursor to the beginning of the stream and
     // (re)reading the header.
     fn seek_to_first_record(&mut self) -> io::Result<bgzf::VirtualPosition> {
-        self.get_mut().seek(bgzf::VirtualPosition::default())?;
+        self.get_mut()
+            .seek_to_virtual_position(bgzf::VirtualPosition::default())?;
+
         self.read_header()?;
+
         Ok(self.get_ref().virtual_position())
     }
 
@@ -369,7 +372,7 @@ where
         I: BinningIndex,
     {
         if let Some(pos) = index.last_first_record_start_position() {
-            self.get_mut().seek(pos)?;
+            self.get_mut().seek_to_virtual_position(pos)?;
         } else {
             self.seek_to_first_record()?;
         }