+++ noodles-vcf/src/io/indexed_reader.rs
@@ -4,7 +4,7 @@ mod builder;
 
 pub use self::builder::Builder;
 
-use std::io::{self, Read, Seek};
+use std::io::{self, BufRead, Read};
 
 use noodles_bgzf as bgzf;
 use noodles_core::Region;
@@ -18,42 +18,31 @@ use crate::{variant::RecordBuf, Header, Record};
 
 /// An indexed VCF reader.
 pub struct IndexedReader<R> {
-    inner: Reader<bgzf::Reader<R>>,
+    inner: Reader<R>,
     index: Box<dyn BinningIndex>,
 }
 
 impl<R> IndexedReader<R> {
     /// Returns a reference to the underlying reader.
-    pub fn get_ref(&self) -> &bgzf::Reader<R> {
+    pub fn get_ref(&self) -> &R {
         self.inner.get_ref()
     }
 
     /// Returns a mutable reference to the underlying reader.
-    pub fn get_mut(&mut self) -> &mut bgzf::Reader<R> {
+    pub fn get_mut(&mut self) -> &mut R {
         self.inner.get_mut()
     }
 
     /// Returns the underlying reader.
-    pub fn into_inner(self) -> bgzf::Reader<R> {
+    pub fn into_inner(self) -> R {
         self.inner.into_inner()
     }
 }
 
 impl<R> IndexedReader<R>
 where
-    R: Read,
+    R: BufRead,
 {
-    /// Creates an indexed VCF reader.
-    pub fn new<I>(inner: R, index: I) -> Self
-    where
-        I: BinningIndex + 'static,
-    {
-        Self {
-            inner: Reader::new(bgzf::Reader::new(inner)),
-            index: Box::new(index),
-        }
-    }
-
     /// Reads the VCF header.
     pub fn read_header(&mut self) -> io::Result<Header> {
         self.inner.read_header()
@@ -69,10 +58,7 @@ where
     }
 
     /// Returns an iterator over records starting from the current stream position.
-    pub fn record_bufs<'r, 'h: 'r>(
-        &'r mut self,
-        header: &'h Header,
-    ) -> RecordBufs<'r, 'h, bgzf::Reader<R>> {
+    pub fn record_bufs<'r, 'h: 'r>(&'r mut self, header: &'h Header) -> RecordBufs<'r, 'h, R> {
         self.inner.record_bufs(header)
     }
 
@@ -94,7 +80,7 @@ where
 
 impl<R> IndexedReader<R>
 where
-    R: Read + Seek,
+    R: bgzf::io::BufRead + bgzf::io::Seek,
 {
     /// Returns an iterator over records that intersects the given region.
     pub fn query<'r, 'h>(
@@ -103,5 +89,21 @@ where
         region: &Region,
     ) -> io::Result<Query<'r, 'h, R>> {
         self.inner.query(header, &self.index, region)
+    }
+}
+
+impl<R> IndexedReader<bgzf::Reader<R>>
+where
+    R: Read,
+{
+    /// Creates an indexed VCF reader.
+    pub fn new<I>(inner: R, index: I) -> Self
+    where
+        I: BinningIndex + 'static,
+    {
+        Self {
+            inner: Reader::new(bgzf::Reader::new(inner)),
+            index: Box::new(index),
+        }
     }
 }